Update pc.css - !important
Discord changed some style settings. This affected paddings and margins in the code. Adding !important seems to fix this issue but a better method will need to take effect to prevent this from happening again.

Possibly adding `!important` to all the new code to prevent it being overwritten later on.